Section entry parsing to Error instead of Panic (#1)
Parsing a section entry is done through `from_key_value_pair` (previously `new`) and would panic if the format of a line was not as expected. This function has been updated to return a Result object so that a failure in parsing a section entry line can be handled depending on the user. This change allows for an additional config, `no_fail_mode`, which will simply log a warning when an entry could not be parsed, rather than panicking.