Do not lazy parse (#2)
Previously, the parser would only lazily parse section entries when entries in that particular section were requested. While this was faster, it is not the experience I want the user to have, as the `parse()` command would succeed, even if there were issues with individual lines in the config file.

This commit updates the parser to parse all sections to ensure they meet the requirements for the given section, but also serializes them to store in an index map as byte arrays. When `get_section_entries` is called, it now deserializes the already-parsed section entries rather than parsing them.