add tests that parse edk2 (#5)
Adds a new test file for parsing edk2 itself. This test is to ensure
any changes do not any config parser as a whole when run against edk2.

This not does mean the parsers continues to be correct. Other tests are
necessary for that. This only means that any changes do not break the
ability to parse in general.

This is a work in progress as only the inf parser is working